インターフェイスを別プロジェクトに分離
