+++ BIDSSMemLib_Mac.sln
@@ -27,6 +27,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TR.SMemIF.Mock", "TR.SMemIF
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TR.SMemIF.Tests", "TR.SMemIF.Tests\TR.SMemIF.Tests.csproj", "{D4444E55-9918-4905-9857-3BC34C7EFEE5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TR.ISMemCtrler", "TR.ISMemCtrler\TR.ISMemCtrler.csproj", "{B44CA6E0-B092-4FBA-9C30-EC6576733929}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TR.ISMemIF", "TR.ISMemIF\TR.ISMemIF.csproj", "{51C50917-9A27-440E-84C7-EFF691B79840}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -181,6 +185,30 @@ Global
 		{D4444E55-9918-4905-9857-3BC34C7EFEE5}.Debug|x86.Build.0 = Debug|Any CPU
 		{D4444E55-9918-4905-9857-3BC34C7EFEE5}.Release|x86.ActiveCfg = Release|Any CPU
 		{D4444E55-9918-4905-9857-3BC34C7EFEE5}.Release|x86.Build.0 = Release|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Debug|x64.Build.0 = Debug|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Release|x64.ActiveCfg = Release|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Release|x64.Build.0 = Release|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Debug|x86.Build.0 = Debug|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Release|x86.ActiveCfg = Release|Any CPU
+		{B44CA6E0-B092-4FBA-9C30-EC6576733929}.Release|x86.Build.0 = Release|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Release|Any CPU.Build.0 = Release|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Debug|x64.Build.0 = Debug|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Release|x64.ActiveCfg = Release|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Release|x64.Build.0 = Release|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Debug|x86.Build.0 = Debug|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Release|x86.ActiveCfg = Release|Any CPU
+		{51C50917-9A27-440E-84C7-EFF691B79840}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
