+++ TR.ISMemCtrler/IContainsSMemIF.cs
@@ -1,0 +1,8 @@
+﻿namespace TR;
+
+/// <summary>SMemIF型インスタンスへのアクセスを提供するインターフェイス</summary>
+public interface IContainsSMemIF
+{
+	/// <summary>SMemIF型プロパティ</summary>
+	ISMemIF? MMF { get; }
+}
