+++ TR.ISMemIF/IRWSemap.cs
