+++ TR.ISMemIF/ISMemIF.cs
